fix: move civilian type icon under features for repo lint (#4140)
Relocate CivilianUnitTypeIcon from app/lib/widgets to the civilian
features folder so it can import map-theme caches without violating
repo.app_widget_imports.
