+++ app/lib/features/game/widgets/civilian/civilian_unit_type_icon.dart
